fix(desktop): native GTK Alt-menu toggle on Linux, modularize main.rs
The Alt-key menu-bar toggle relied on a JS/DOM key listener re-injected on
every page load, so it silently missed the key whenever the webview's page
content didn't have focus. Move it to a native GTK key-press/key-release
listener on the window's toplevel instead -- the same event path that
already makes menu accelerators like CmdOrCtrl+N fire reliably regardless
of focus.

While in main.rs, also split the 1400+ line monolith into config/debug_log/
menu/window/commands/alt_menu modules, dedupe the two drifted window-builder
code paths, make ConfigState's locks poison-safe, store menu-checkbox
handles instead of re-walking the menu tree by string ID, and replace
silent `let _ = ...` error-swallowing with logged errors.
