+++ desktop/src-tauri/Cargo.toml
@@ -21,6 +21,9 @@ window-vibrancy = "0.7.1"
 url = "2.5"
 reqwest = { version = "0.13", default-features = false, features = ["rustls"] }
 
+[target.'cfg(target_os = "linux")'.dependencies]
+gtk = "0.18"
+
 [features]
 default = ["custom-protocol"]
 custom-protocol = ["tauri/custom-protocol"]
